+++ demo/demo1/types.ts
@@ -1,3 +1,5 @@
+import { ScopeInf } from '../../src/GlobalState/GlobalState';
+
 export interface UserScopeInf {
     name: string;
     city: string;